Fix make-bootable w/ "docker save".
"docker save" doesn't name any of the layer related files with the hash
of the tarfile contents for that layer, like "podman save" does.

Instead we extract to a file with a temporary name, sha256 the tar
file as we extract it, and then rename at the end with the measured
diff_id.